Relatively major restructure: - Replaced "init" and "make" functions with type constructors (eliminates "pre-processing") - Reorganized "problem-types" file into several smaller groups - Precomputing x- and y- offsets for calculating rotational part of base flux - Added rotational offset for MovingGrid
Still working on a few validation tasks, but I think it's getting close
