+++ benchmarks/cyl_rot/cyl_rot2.jl
@@ -4,45 +4,18 @@ Rotating cylinder benchmark 2:
 Body-fixed frame
 
 Expected results:
-    Converge to CD=2.1562, CL=-0.17185 by T=100
 """
-
-include("../../src/ibpm.jl")
-using .ibpm
 using MAT
+include("config.jl")  # Set up grid and common variables
 
-# Define grid
-nx = 200  # num of x points on finest domain
-ny = 200  # num of y points on finest domain (length in y-dirn is len/m * n )
-mg = 5   # num domains
-offx = 2.0; # offset in x dirn (on fine domain, x-grid runs from -offx to len-offx.
-offy = 2.0; # offset in y dirn (same as offx but in y-dirn)
-
-len = 4.0  # length of domain in x-direction
-
-# Other parameters
-Re = 20.0
-Δt = 1e-2
-
-# Initialize grid
-grid = ibpm.make_grid(nx, ny, offx, offy, len, mg=mg)
-
-# Set up motion
-Uinf(t) = 1.0;
-#θ̇(t) = 0.1;
-#θ(t) = θ̇(t)*t;
-
-# For debugging
-t_init = 0;  # Let flow equilibrate first
-θ̇(t) = t>t_init ? 0.1 : 0.0
-θ(t) = θ̇(t)*(t-t_init);
-motion = ibpm.MovingGrid(Uinf, θ, θ̇)
+motion = ibpm.MovingGrid(t -> Uinf,
+						 t -> \Omega*t,
+						 t -> \Omega)
 
 # Create cylinder
-r = 0.5; # Cylinder radius
-cyls = [ibpm.make_cylinder( r, grid.h, 0.0, 0.0, motion; n=78 )]
+cyls = [ibpm.make_cylinder( r, grid.h, x0, y0; motion=motion, n=nb )]
 
-prob = ibpm.init_prob(grid, cyls, Re, Δt);
+prob = ibpm.init_prob(grid, cyls, Δt, Re);
 state = ibpm.init_state(prob);
 ibpm.base_flux!(state, prob, 0.0)  # Initialize irrotational base flux
