+++ benchmarks/eldredge/pitch_maneuver.jl
@@ -22,7 +22,7 @@ motion = ibpm.MovingGrid(t -> Uinf, θ, θ̇)
 
 # Create plate
 bodies = [ibpm.make_plate( L, α, grid.h, x0, y0; motion=motion, n=nb)]
-prob = ibpm.init_prob(grid, bodies, Δt, Re, Uinf=Uinf);
+prob = ibpm.IBProblem(grid, bodies, Δt, Re, Uinf=Uinf);
 
 # Load steady boundary layer solution from plate_steady.jl
 @load "benchmarks/eldredge/results/steady.bson" state
