+++ examples/osc_plate.jl
@@ -27,8 +27,8 @@ L = 1.0; # Plate length
 x0, y0 = 0.0, 0.5
 bodies = [ibpm.make_plate( L, α, grid.h, x0, y0, motion; n=51 )]
 
-prob = ibpm.init_prob(grid, bodies, Δt, Re);
-state = ibpm.init_state(prob);
+prob = ibpm.IBProblem(grid, bodies, Δt, Re);
+state = ibpm.IBstate(prob);
 ibpm.base_flux!(state, prob, 0.0)  # Initialize irrotational base flux
 
 T=2.0*2π
